Weights type: uniformly import as WeightT
__Context__
The `Weights` type is currently imported as either `Weights` or `WeightsT`.  The latter case is
used when the type conflicts with the `core/weights.js` module which is imported as `Weights` as
well.

__Advantages__
This changes all instances of importing the type `Weights` to import as `WeightsT` so we have
consistency, which has the following advantage:

(1) Searching for instances of the `Weights` type is much simpler as we don't need to also sort through instances of importing the `core/weights.js` module (and vice versa).

__Other Considerations__
If the `Weights` type is uniformly imported as `WeightsT`, we might consider
defining the type as `WeightsT`, as the reader would not know on face value how that type was
imported elsewhere.  It's not the cleanest type signature, but worth a thought.

__Test Plan__
Trivial naming change, tests pass.